requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.0.0
idna==2.10
ipaddress==1.0.23
lxml==4.6.2
ply==3.11
pyasn1==0.4.8
pycryptodomex==3.9.9
pysmi==0.3.4
pysnmp==4.4.12
pytz==2020.5
requests==2.25.1
route53==1.0.1
urllib3==1.26.2
